backend/requirements.txt
amqp==5.3.1
billiard==4.2.4
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
tzdata==2025.3
tzlocal==5.3.1
uuid7==0.1.0
uuid_utils==0.12.0
vine==5.1.0
wcwidth==0.2.14
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.127.0
FastAPI-SQLAlchemy==0.2.1
greenlet==3.3.0
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
celery[redis]
redis
pyjwt[crypto]
passlib[bcrypt]
pydantic-settings
bcrypt
websockets
aioboto3
python-multipart